Refactored SQL sync code for postgres implementation, green test suite
