+++ doltpy/etl/sql_sync/__init__.py
@@ -1,4 +1,4 @@
-from .tools import sync_from_dolt, sync_to_dolt
+from .sync_tools import sync_from_dolt, sync_to_dolt
 from .dolt import (get_source_reader as get_dolt_source_reader,
                    get_table_reader as get_dolt_table_reader,
                    get_table_reader_diffs as get_dolt_table_reader_diffs,
