+++ doltpy/etl/sql_sync/postgres.py
@@ -1,26 +1,29 @@
-from typing import List, Tuple, Mapping, Iterable, Any, Callable
-from doltpy.etl.sql_sync.tools import (DoltAsSourceWriter, DoltAsTargetReader, TableMetadata, Column, TableUpdate)
+from typing import List, Any, Callable
+from doltpy.etl.sql_sync.db_tools import (TableMetadata,
+                                          Column,
+                                          DoltAsSourceWriter,
+                                          DoltAsTargetReader,
+                                          TableUpdate,
+                                          build_target_writer,
+                                          build_source_reader,
+                                          get_table_reader,
+                                          get_insertion_lists)
 import logging
 
 logger = logging.getLogger(__name__)
 
 
 def get_target_writer(conn, update_on_duplicate: bool = True) -> DoltAsSourceWriter:
-    def inner(table_data_map: Mapping[str, Iterable[tuple]]):
-        for table, table_update in table_data_map.items():
-            table_metadata = get_table_metadata(table, conn)
-            pks_to_drop, data = table_update
-            pks_to_drop = list(pks_to_drop)
-            logger.info('Dropping {} primary keys from {}'.format(len(pks_to_drop), table))
-            drop_primary_keys(conn, table_metadata, pks_to_drop)
-            data = list(data)
-            logger.info('Writing {} rows to table {}'.format(len(data), table))
-            write_to_table(table_metadata, conn, data, update_on_duplicate)
+    """
 
-    return inner
+    :param conn:
+    :param update_on_duplicate:
+    :return:
+    """
+    return build_target_writer(conn, get_table_metadata, get_insert_query, update_on_duplicate)
 
 
-def get_source_reader(conn, schema: str, reader: Callable[[str, Any], TableUpdate]) -> DoltAsTargetReader:
+def get_source_reader(conn, reader: Callable[[str, Any], TableUpdate] = None) -> DoltAsTargetReader:
     """
     Given a connection and a reader provides a function that turns a set of tables in to a data structure containing
     the contents of each of the tables.
@@ -29,42 +32,8 @@ def get_source_reader(conn, schema: str, reader: Callable[[str, Any], TableUpdat
     :param reader:
     :return:
     """
-    def inner(tables: List[str]):
-        result = {}
-        database_tables = get_tables(conn, schema)
-        missing_tables = [table for table in tables if table not in database_tables]
-        if missing_tables:
-            logger.error('The following tables are missign, exiting:\n{}'.format(missing_tables))
-            raise ValueError('Missing tables {}'.format(missing_tables))
-
-        for table in tables:
-            logger.info('Reading tables {}'.format(table))
-            result[table] = reader(table, conn)
-
-        return result
-
-    return inner
-
-
-def get_table_reader():
-    """
-    When syncing from a relational database such as MySQL the database has only a single concept of state, that is the
-    current state. We simply capture this state by reading out all the data in the database.
-    :return:
-    """
-    def inner(table_name: str, conn: Any):
-        table_metadata = get_table_metadata(table_name, conn)
-        query = '''
-            SELECT
-                {cols}
-            FROM
-                {table_name}
-        '''.format(cols=','.join(col.col_name for col in table_metadata.columns), table_name=table_name)
-        cursor = conn.cursor()
-        cursor.execute(query)
-        return [tup for tup in cursor]
-
-    return inner
+    reader_function = reader or get_table_reader()
+    return build_source_reader(conn, get_table_metadata, reader_function)
 
 
 def get_tables(conn, schema: str = None):
@@ -80,14 +49,6 @@ def get_tables(conn, schema: str = None):
     cursor = conn.cursor()
     cursor.execute(query)
     return [tup[0] for tup in cursor]
-
-
-def write_to_table(table_metadata: TableMetadata, conn, data: List[tuple], update_on_duplicate: bool = True):
-    insert_query = get_insert_query(table_metadata, update_on_duplicate)
-    cursor = conn.cursor()
-    cursor.executemany(insert_query, data)
-    conn.commit()
-    cursor.close()
 
 
 def get_table_metadata(table_name: str, conn) -> TableMetadata:
@@ -124,43 +85,6 @@ def _get_primary_key_cols(table_name: str, conn) -> List[str]:
     return [tup[0] for tup in cursor]
 
 
-def drop_primary_keys(conn, table_metadata: TableMetadata, primary_key_values: List[tuple]):
-    """
-    Drops a given list of primary keys from the database represented by the conn parameter.
-    :param conn:
-    :param table_metadata:
-    :param primary_key_values:
-    :return:
-    """
-    if not primary_key_values:
-        return
-    pks = [col.col_name for col in table_metadata.columns if col.key]
-    query_template = '''
-        DELETE FROM 
-            {table_name}
-        WHERE
-            {delete_filter}
-    '''
-
-    delete_filter = get_filters(pks)
-
-    query = query_template.format(table_name=table_metadata.name, delete_filter=delete_filter)
-    cursor = conn.cursor()
-    cursor.executemany(query, primary_key_values)
-    conn.commit()
-
-
-def get_filters(cols: List[str]):
-    if len(cols) == 1:
-        delete_clause = '{col} = %s'.format(col=cols[0])
-    else:
-        base_delete_clause = '{first_col} = %s AND {rest_cols}'
-        rest_cols = 'AND '.join(['{} = %s'.format(col) for col in cols[1:]])
-        delete_clause = base_delete_clause.format(first_col=cols[0], rest_cols=rest_cols)
-
-    return delete_clause
-
-
 def get_insert_query(table_metadata: TableMetadata, update_on_duplicate: bool = True) -> str:
     col_list, wildcard_list = get_insertion_lists(table_metadata)
     base_query = '''
@@ -181,12 +105,3 @@ def get_insert_query(table_metadata: TableMetadata, update_on_duplicate: bool = 
     else:
         return base_query
 
-
-def get_insertion_lists(table_metadata: TableMetadata) -> Tuple[List[str], List[str]]:
-    col_list, wildcard_list = [], []
-
-    for col in table_metadata.columns:
-        col_list.append(col.col_name)
-        wildcard_list.append('%s')
-
-    return col_list, wildcard_list
