+++ doltpy/etl/sql_sync/tests/test_dolt.py
@@ -12,11 +12,8 @@ from doltpy.etl.sql_sync.tests.helpers.data_helper import (TEST_TABLE_COLUMNS,
                                                            TEST_DATA_APPEND_MULTIPLE_ROWS,
                                                            TEST_DATA_APPEND_SINGLE_ROW,
                                                            TEST_DATA_UPDATE_SINGLE_ROW)
-from doltpy.etl.sql_sync.dolt import (get_table_reader_diffs,
-                                      get_table_reader,
-                                      get_table_metadata,
-                                      write_to_table,
-                                      get_target_writer)
+from doltpy.etl.sql_sync.dolt import get_table_reader_diffs, get_table_reader, get_target_writer
+from doltpy.etl.sql_sync.mysql import get_table_metadata
 from doltpy.core.dolt import Dolt
 import logging
 
@@ -107,11 +104,6 @@ def _dolt_table_read_helper(repo: Dolt, table_name: str):
     result = [tup for tup in cursor]
     conn.close()
     return result
-
-
-def test_drop_primary_keys():
-    # test that given new data appropriate keys will be dropped
-    pass
 
 
 def test_get_table_metadata(create_dolt_test_data_commits):
